Make a function to savefig in cross-platform way
The current calls to plt.savefig only work in Windows. I made a
`savefig` function in `plot_functions` that uses the `os` module to make
the file path in a cross-platform way. I replaced the calls to
`plt.savefig` in the two notebooks that use it. The figures remained the
same so I'm not commiting them as well.